Bug#25943754: FIX GCC 7 COMPILE WARNINGS
Followup patch: developerstudio12.5 incorrectly complains about
line 325: Error: Overloading ambiguity between
"built-in operator==(int, float)" and "built-in operator==(int, double)".

Fix: explicit cast to int

+++ storage/innobase/trx/trx0rseg.cc
@@ -320,7 +320,8 @@ trx_rseg_create(
 		|| rseg_id < srv_tmp_rollback_segments);
 	ut_ad(fsp_is_system_temporary(space_id)
 		|| rseg_id < srv_rollback_segments);
-	ut_ad(to_int(space->purpose) == (fsp_is_system_temporary(space_id)
+	ut_ad(to_int(space->purpose)
+		== static_cast<int>(fsp_is_system_temporary(space_id)
 				? FIL_TYPE_TEMPORARY
 				: FIL_TYPE_TABLESPACE));
 	ut_ad(univ_page_size.equals_to(page_size_t(space->flags)));
